Conflicts:
	Crocante/Scripts y Datos/PAsCrocante.sql
	Crocante/src/main/java/com/unal/crocante/MenuCliente.java

+++ Crocante/src/main/java/com/unal/crocante/venta/pedido/AgregarPedido.java
@@ -203,7 +203,7 @@ public class AgregarPedido extends javax.swing.JFrame {
             s = conexion.prepareStatement(queryGetCc);
             ResultSet result = s.executeQuery();
             result.next();
-            cc = result.getInt(1);
+            cc = result.getLong(1);
         } catch (SQLException ex) {
             Logger.getLogger(VentaPrincipal.class.getName()).log(Level.SEVERE, null, ex);
         }
